requirements.txt
asgiref==3.11.0
Django==3.2.7
django-filter==25.2
pytz==2025.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
